requirements.txt
google-generativeai>=0.8.0
langchain>=0.3.0
langchain-google-genai>=2.0.0
pinecone>=5.0.0
langchain-pinecone>=0.2.0
flask>=3.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
sentence-transformers>=3.0.0